requirements.txt
torch
torch.nn
torch.nn
numpy
matplotlib.pyplot 